Finished Project

+++ src/components/Country.jsx
@@ -1,14 +1,12 @@
 import React from 'react';
 import Medal from './Medal';
-import { TrashFill } from 'react-bootstrap-icons';
-import {DatabaseFillAdd} from 'react-bootstrap-icons';
-import {ArrowCounterclockwise} from 'react-bootstrap-icons';
+import { TrashFill, Save, ArrowCounterclockwise } from 'react-bootstrap-icons';
 import Card from 'react-bootstrap/Card';
 import Badge from 'react-bootstrap/Badge';
 import ListGroup from 'react-bootstrap/ListGroup';
 
 const Country = (props) => {
-  const { country, medals, onIncrement, onDecrement, onDelete, onSave, onReset } = props;
+  const { country, medals, onIncrement, onDecrement, onDelete, onSave, onReset, canDelete, canPatch } = props;
   const getMedalsTotal = (country, medals) => {
     let sum = 0;
     // use medal count displayed in the web page for medal count totals
@@ -39,11 +37,11 @@ const Country = (props) => {
           { renderSaveButton() ?
             <React.Fragment>
               {/* TODO: use Bootstrap stying / icons */}
-              <DatabaseFillAdd onClick={ () => onSave(country.id) }>save</DatabaseFillAdd>
-              <ArrowCounterclockwise onClick={ () => onReset(country.id) }>reset</ArrowCounterclockwise>
+              <ArrowCounterclockwise onClick={ () => onReset(country.id) } className='icon-btn' />
+              <Save onClick={ () => onSave(country.id) } className='icon-btn' />
             </React.Fragment>
             :
-            <TrashFill onClick={() => onDelete(country.id)} className='icon-btn' style={{ color:'red' }} />
+            canDelete && <TrashFill onClick={() => onDelete(country.id)} className='icon-btn' style={{ color:'red' }} />
           }
         </Card.Title>
         <ListGroup variant="flush">
@@ -52,6 +50,7 @@ const Country = (props) => {
           <Medal  
             country={ country } 
             medal={ medal } 
+            canPatch={ canPatch }
             onIncrement={ onIncrement } 
             onDecrement={ onDecrement } />
         </ListGroup.Item>
